+++ controls/popups/src/common/resize.ts
@@ -405,12 +405,18 @@ function resizeNorth(e: MouseEvent | TouchEvent): void {
     let boundaryRectValues: ClientRect;
     const pageY: number = (getEventType(e.type) === 'mouse') ? (e as MouseEvent).pageY : (e as TouchEvent).touches[0].pageY;
     const targetRectValues: ClientRect = getClientRectValues(targetElement);
+    const borderValue: number = isNOU(containerElement) ? 0 : containerElement.offsetHeight - containerElement.clientHeight;
     if (!isNOU(containerElement)) {
         boundaryRectValues = getClientRectValues(containerElement);
     }
     if (!isNOU(containerElement) && (targetRectValues.top - boundaryRectValues.top) > 0 ) {
         calculateValue = true;
     } else if (isNOU(containerElement) && pageY > 0) {
+        calculateValue = true;
+    }
+    else if (!isNOU(containerElement) &&
+    (Math.floor((targetRectValues.top - boundaryRectValues.top) + targetRectValues.height +
+    (boundaryRectValues.bottom - targetRectValues.bottom)) - borderValue) <= maxHeight) {
         calculateValue = true;
     }
     const currentHeight: number = originalHeight - (pageY - originalMouseY);
@@ -423,9 +429,7 @@ function resizeNorth(e: MouseEvent | TouchEvent): void {
             let top: number = (originalY - containerTop) + (pageY - originalMouseY);
             top = top > 0 ? top : 1;
             targetElement.style.height = currentHeight  + 'px';
-            if (!(targetElement.classList.contains('e-dlg-modal') && (targetElement.style.top === '0px'))) {
-                targetElement.style.top = top + 'px';
-            }
+            targetElement.style.top = top + 'px';
         }
     }
 }
@@ -481,9 +485,7 @@ function resizeWest(e: MouseEvent | TouchEvent): void {
                 targetElement.style.width = calculatedWidth + 'px';
             }
             if (setLeft) {
-                if (!(targetElement.classList.contains('e-dlg-modal') && (targetElement.style.left === '0px'))) {
-                    targetElement.style.left = left + 'px';
-                }
+                targetElement.style.left = left + 'px';
                 if (left === 1) {
                     setWidth = false;
                 } else {