+++ controls/popups/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+## 27.1.56 (2024-10-23)
+
+### Dialog
+
+#### Bug Fixes
+
+- `#I633873` - Resolved the inconsistent resizing behaviour observed with the modal dialog.
+
 ## 27.1.53 (2024-10-15)
 
 ### Dialog